Add class for the maze agent
The maze solver will need the ability to internally represent the maze data
matrix, navigate that matrix, and save its findings. These tasks can be
accomplished by an "agent" object that can navigate and save the solution for
a maze.

This commit adds the class MazeAgent, which will contain all the data and
methods necessary to perform the tasks described above.

Currently it only initializes the object withe the maze data.