Initialize MazeAgent internal maze representation
The MazeAgent needs an internal representation of where it has explored in
the maze so far. This will be handled by another matrix of the same
dimensions as the original maze.

This matrix will be all 0s initially, and then a "1" will be entered at the
coordinates of the maze entry. Each explored step in the maze will be
represented by an integer in the corresponding matrix entry. This integer
will be the number of steps back to the maze start.