Move find_starting_coordinates to the MazeAgent class
The find_starting_coordinates method was added before the MazeAgent class was
created. But its functionality really belongs in the MazeAgent class, as the
class should be able to simply take a matrix of maze data and process it.

This commit moves the method to the MazeAgent class. It also updates the
MazeAgent __init__ to call self.find_starting_coordinates() to get the maze
start coordinates instead of taking them as an argument.

This commit also adds a maze_start property to the MazeAgent class that returns
the starting coordinates as a tuple for comparison/display purposes.