Add method to determine if a given location is an exit.
The MazeAgent needs a way to determine if a given location is an exit or not
in order to actually solve the maze.

We define a maze exit here as a valid(non-wall) coordinate that is in the final
row of the matrix.

This commit adds the method agent_at_exit() to MazeAgent that takes a location
coordinate as a list, and returns True if it is at a valid exit.