US2: Add method to find the exit in a straight line.
User Story 2: This commit should allow the agent to complete the user story 2
maze. To verify, run the following tests:
```
python -m unittest tests.test_user_story_integration.test_user_story_2
```

The main functionality of the MazeAgent is to take a matrix of maze data and
find a path to the maze exit. So it needs a way to internally navigate/process
the maze data.

This commit adds the MazeAgent method find_exit(). This method can be called
after initializing the maze agent, and it traverses the maze in a straight
line until it finds the exit, marking down each spot in the agent's route
matrix as it goes.

The longest chain of consecutive integers in the routes matrix represents the
path from the maze start to the maze exit.

This is only the initial implementation of the find_exit method, so it can only
move in a straight line to the exit. As a result, the method will be unable to
find the it in any maze that is more complicated than a straight line.

So this commit also implements a timeout in the find_exit method. The timeout
is a simple integer counter in the while loop that breaks out of the loop if
the counter hits the timeout value (1000 by default).

Most of the mazes used by this solver are simple enough that this timeout should
only trigger if the agent gets stuck.