Save maze exit coords as attributes of MazeAgent
Previously, the find_exit() method would simply return the coordinates of
the maze exit. But the MazeAgent would benefit from saving the coordinates on
its current object as well.

So this commit adds the maze exit row and column as attributes (like the maze
start coordinates) and a computed property that formats the maze exit
coordinates together in a tuple.