US3: Add ability to turn agent right in find_exit
User Story 3: This commit should allow the agent to complete the user story 3
maze. To verify, run the following tests:
```
python -m unittest tests.test_user_story_integration.test_user_story_3
```

Previously, the MazeAgent.find_exit() method was only able to check and advance
to the space directly below it. To solve more complicated mazes, the agent must
be able to turn as well.

This commit adds the ability for the agent to advance to the right of its
current location in find_exit().