Add ability for agent to move Left and Up
Previously, the MazeAgent had the ability to evaluate and move to spaces
below, and to the right of its current position via the find_exit() method.
But the agent will also need move left and up if it is to fully navigate and
solve more copmplex mazes.

This commit adds cases to the MazeAgent.find_exit() method that allow the agent
to evaulate and traverse to spaces to the left and above the current space.