US4: Add more criteria to evaluate_coord()
User Story 4: This commit should allow the agent to complete the user story 4
maze. To verify, run the following tests:
```
python -m unittest tests.test_user_story_integration.test_user_story_4
```

Previously, the agent was able to navigate in all directions, but it could
not differentiate between spaces it had visited and those it had not. This
led to situations where the agent could get stuck going around corners.

One solution to this problem is to have the agent check if it has been to a
space already when it is evaluating it. This will help the agent not navigate
backwards inadvertently.