Add path tracking to MazeAgent
Previously, the maze path data from start to finish was saved in the MazeAgent
routes matrix. But instead of parsing the routes matrix for path traveled,
we can save the path data in a list of coordinates as we traverse the maze.

This commit adds the code to track the coordinates traversed as the maze is
navigated.