Display maze with solution path
Previously there was a utility method to display a given matrix in ASCII
format, but it couldn't display the maze solution within the maze.

This commit updates this method to optionally take a list of coordinates
(which represent the path traversed by the agent) and display them within the
maze graphical output (agent path represented by xs).