Add MazeAgent.evaluate_coord tests for explored spaces
This commit adds tests for the MazeAgent.evaluate_coord method to ensure it
works correctly with all types of already explored spaces (the corresponding
entry in maze_agent.routes is greater than 0).

The maze_agent.routes entry can be less than, equal to, or greater than the
current step. MazeAgent.evaluate_coord() only returns True if the cooresponding
routes value is equal to the current step (which happens in practice as the
agent moves back along its path).