Add print_maze_state method to MazeAgent class
This commit adds a class method to the MazeAgent class to print the agent's
maze with all the path info it currently has.

If the maze has been initialized with the agent but not solved, it will print
just the maze. If the maze has been solved, it will also print the solution path
and all other visited spaces.